+++ src/components/button-add/button-add.component.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { Button, ButtonColors } from "./button";
-import PlusIcon from "../assets/icons/plus_icon.svg";
+import { Button, ButtonColors } from "../button";
+import PlusIcon from "../../assets/icons/plus_icon.svg";
 import styles from "./button-add.module.scss";
 
 interface IProps {
