+++ src/components/button-add/index.ts
@@ -1,0 +1,1 @@
+export { ButtonAdd } from "./button-add.component";
