+++ src/components/index.ts
@@ -4,3 +4,11 @@ export { Overlay } from "./overlay";
 export { Profile } from "./profile";
 export { Avatar } from "./avatar";
 export { ButtonColors, Button } from "./button";
+export { ButtonAdd } from "./button-add";
+export { Footer } from "./footer";
+export { InputNumber } from "./input-number";
+export { InputPrice } from "./input-price";
+export { InputText } from "./input-text";
+export { Price } from "./price";
+export { Select } from "./select";
+export type { SelectOption } from "./select";
