+++ src/components/input-number/index.ts
@@ -1,0 +1,1 @@
+export { InputNumber } from "./input-number.component";
