+++ src/components/input-number/input-number.module.scss
@@ -1,4 +1,4 @@
-@use "../styles/typography";
+@use "../../styles/typography";
 
 .input {
   @include typography.light;
