+++ src/components/input-price/index.ts
@@ -1,0 +1,1 @@
+export { InputPrice } from "./input-price.component";
