+++ src/components/input-text/index.ts
@@ -1,0 +1,1 @@
+export { InputText } from "./input-text.component";
