+++ src/components/input-text/input-text.module.scss
@@ -1,0 +1,30 @@
+@use "../../styles/typography";
+
+.input {
+  @include typography.light;
+
+  label {
+    @include typography.regular;
+
+    width: 100%;
+    box-sizing: border-box;
+    padding-left: 0.5rem;
+  }
+
+  input {
+    @include typography.light;
+
+    width: 100%;
+    margin-top: 0.5rem;
+    border-radius: 100px;
+
+    border: 2px solid var(--primary);
+    padding: 1rem;
+    box-sizing: border-box;
+    max-width: 100%;
+
+    &:focus {
+      outline: none;
+    }
+  }
+}
