+++ src/components/price/index.ts
@@ -1,0 +1,1 @@
+export { Price } from "./price.component";
