+++ src/components/select/index.ts
@@ -1,0 +1,2 @@
+export { Select } from "./select.component";
+export type { SelectOption } from "./select.component";
