+++ src/components/select/select.module.scss
@@ -1,4 +1,4 @@
-@use "../styles/typography";
+@use "../../styles/typography";
 
 .select {
   @include typography.light;
