+++ src/product-list/product-list-item-toggleable.component.tsx
@@ -1,6 +1,6 @@
 import type { FC } from "react";
 import React from "react";
-import type { IProduct } from "../product/product.interface";
+import type { IProduct } from "../product";
 import { ProductListItem } from "./product-list-item";
 import { ProductListItemToggleButton } from "./product-list-item-toggle-button";
 
@@ -11,22 +11,20 @@ interface IProps {
   onItemSelected: (i: IProduct) => void;
 }
 
-const ProductListItemToggleable: FC<IProps> = (props) => {
-  return (
-    <>
-      {props.items.map((i, index) => {
-        const onToggle = () => props.onToggleItem(i.title);
+const ProductListItemToggleable: FC<IProps> = (props) => (
+  <>
+    {props.items.map((i, index) => {
+      const onToggle = () => props.onToggleItem(i.title);
 
-        const onSelect = () => props.onItemSelected(i);
+      const onSelect = () => props.onItemSelected(i);
 
-        return (
-          <ProductListItem className={props.className} key={i.title} index={index} {...i} onClick={onSelect}>
-            <ProductListItemToggleButton onClick={onToggle} active={i.active} />
-          </ProductListItem>
-        );
-      })}
-    </>
-  );
-};
+      return (
+        <ProductListItem className={props.className} key={i.title} index={index} {...i} onClick={onSelect}>
+          <ProductListItemToggleButton onClick={onToggle} active={i.active} />
+        </ProductListItem>
+      );
+    })}
+  </>
+);
 
 export { ProductListItemToggleable };
