+++ src/product-list/product-list-ul.component.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-const ProductListUl: React.FC = (props) => {
-  return <ul style={{ listStyle: "none", margin: 0, padding: 0 }}>{props.children}</ul>;
-};
+const ProductListUl: React.FC = (props) => (
+  <ul style={{ listStyle: "none", margin: 0, padding: 0 }}>{props.children}</ul>
+);
 
 export { ProductListUl };
 export default ProductListUl;
