+++ src/product-list/product-list.interface.ts
@@ -1,9 +1,10 @@
-import type { IProduct } from "../product/product.interface";
+import type { IProduct } from "../product";
 
 interface IProductList {
   title: string;
   items: IProduct[];
   id: number;
+  lastEditedDate: Date;
 }
 
 export type { IProductList };
