+++ src/product/index.ts
@@ -1,0 +1,3 @@
+export { ProductForm } from "./product-form";
+export { ProductPriorityIcon, ProductPriorityEnum } from "./product-priority-icon";
+export type { IProduct } from "./product.interface";
