+++ src/product/product-form/index.ts
@@ -1,0 +1,1 @@
+export { ProductForm } from "./product-form.component";
