+++ src/product/product-priority-icon/index.ts
@@ -1,0 +1,2 @@
+export { ProductPriorityEnum } from "./product-priority.enum";
+export { ProductPriorityIcon } from "./product-priority-icon.component";
