+++ src/product/product-priority-icon/product-priority-icon.module.scss
